--- Peaje Automatico/src/Entidades/Cola_Comun_Ruta.java
@@ -1,18 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package Entidades;
-
-import java.util.LinkedList;
-import java.util.Queue;
-
-/**
- *
- * @author apicos
- */
-public class Cola_Comun_Ruta {
-
-    public static Queue <Vehiculo> cola;
-}
