+++ Peaje Automatico/src/Entidades/Colas/Cola_Comun_Ruta.java
@@ -3,8 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Entidades;
+package Entidades.Colas;
 
+import Entidades.Vehiculo;
 import java.util.LinkedList;
 import java.util.Queue;
 
@@ -12,7 +13,7 @@ import java.util.Queue;
  *
  * @author apicos
  */
-public class Colas_Vehiculos_Clasificados {
-    public static Queue<Vehiculo> especiales;
-    public static Queue<Vehiculo> normales;
+public class Cola_Comun_Ruta {
+
+    public static Queue <Vehiculo> cola;
 }
