+++ Peaje Automatico/src/Entidades/Colas/Colas_Vehiculos_Clasificados.java
@@ -1,0 +1,19 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package Entidades.Colas;
+
+import Entidades.Vehiculo;
+import java.util.LinkedList;
+import java.util.Queue;
+
+/**
+ *
+ * @author apicos
+ */
+public class Colas_Vehiculos_Clasificados {
+    public static Queue<Vehiculo> especiales;
+    public static Queue<Vehiculo> normales;
+}
