+++ Peaje Automatico/src/Entidades/Proyecto_peaje.java
@@ -5,6 +5,9 @@
  */
 package Entidades;
 
+import Entidades.Cajas.Caja_de_vehiculos;
+import Entidades.Cajas.Caja_de_finalizados;
+
 /**
  *
  * @author Agustín Picos
